requirements.txt
# external requirements
python-dotenv>=0.5.1

# backwards compatibility
pathlib2

# project requirements
selenium
pandas
tqdm
requests